pyproject.toml
[project]
name = "crypto-trading-bot-1"
version = "1.1.0"
description = "A modular crypto trading bot with FastAPI backend, PostgreSQL model registry, ML pipeline, and notifier integration."
requires-python = ">=3.10"
authors = [{ name = "Ovie", email = "oviemsaniyo30@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # --- Core ---
    "fastapi==0.111.0",
    "uvicorn==0.35.0",
    "sqlalchemy==2.0.41",
    "alembic==1.16.4",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
    # --- Trading + Exchange ---
    "python-binance==1.0.29",
    "python-telegram-bot==22.5",
    "ta==0.11.0",
    "backoff==2.2.1",
    # --- Auth ---
    "passlib==1.7.4",
    "bcrypt==3.2.2",
    "python-jose==3.3.0",
    "email-validator==2.2.0",
    "python-multipart==0.0.20",
    # --- ML & Data ---
    "pandas==2.3.0",
    "numpy==1.26.4",
    "scikit-learn==1.3.2",
    "xgboost==3.0.2",
    "imbalanced-learn==0.12.4",
    "mlflow==2.13.0",
    "matplotlib==3.10.7",
    "plotly==6.1.2",
    "joblib==1.5.1",
    # --- Utils ---
    "jinja2==3.1.6",
    "pydantic==2.11.7",
    "pydantic-settings>=2.12.0",
]

[project.optional-dependencies]
dev = [
    # Lint & Format
    "mypy",
    "ruff",

    # Testing
    "pytest",
    "pytest-asyncio",
    "pytest-mock",

    # Type stubs
    "types-requests",
    "types-psycopg2",
    "types-python-jose",
    "types-passlib",
    "pandas-stubs",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# ============================================================
#   🧹 Formatting & Linting
# ============================================================
[tool.ruff]
target-version = "py312"
line-length = 100

exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "build",
    "dist",
    "migrations",
    "*.egg-info",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
# ============================================================
#   🔍 MYPY Configuration
# ============================================================

[tool.mypy]
# Check everything under src + tests
files = ["src", "tests"]
python_version = "3.12"

# Good for large ML + API projects
ignore_missing_imports = true  
strict_optional = true
warn_unused_configs = true
check_untyped_defs = false
disallow_untyped_defs = false
no_implicit_optional = false
# Optional tuning:
# allow_untyped_calls = true
# allow_untyped_defs = false


[[tool.mypy.overrides]]
module = ["binance.*", "sqlalchemy.*", "pydantic.*", "python_binance.*", "streamlit*"]
ignore_missing_imports = true
# ============================================================
#   🧪 Testing Configuration
# ============================================================

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --maxfail=3 --tb=short --cov=src --cov-report=term-missing --cov-report=html --cov-report=xml"
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
filterwarnings = [
    "ignore::DeprecationWarning:sqlalchemy.*",
    "ignore::RuntimeWarning:numpy.*"
]

# ============================================================